Minor clean up changes
Fixed the magnifying glass bug, couple files updated to new swift

+++ SustainabilityAppV1/MarqueeLabel.m
@@ -1171,7 +1171,7 @@ CGPoint MLOffsetCGPoint(CGPoint point, CGFloat offset);
     }
     
     // Do not allow negative values
-    _continuousMarqueeExtraBuffer = fabsf(continuousMarqueeExtraBuffer);
+    _continuousMarqueeExtraBuffer = fabs(continuousMarqueeExtraBuffer);
     [self updateSublabelAndLocations];
 }
